refact: filename processing
This commit updates the ExCnab dependency version to "~> 1.3.4" to
incorporate the latest changes and improvements. Additionally, the
`run/2` function in the `ExCnab.Cnab240.Services.GetFileInfo` module has
been refactored to simplify the filename processing logic. The changes
involve modifying the template for extracting information from the
filename and adjusting the positions accordingly. The improvements
enhance the accuracy and reliability of the extracted information. No
known issues or consequences are expected with this change.

Fixes: #123